Fix types passed to ride_type_has_flag

+++ src/openrct2/actions/RideSetPriceAction.hpp
@@ -59,21 +59,21 @@ public:
 
         if (_rideIndex >= MAX_RIDES || _rideIndex == RIDE_ID_NULL)
         {
-            log_warning("Invalid game command for ride %u", _rideIndex);
+            log_warning("Invalid game command for ride %u", uint32_t(_rideIndex));
             return MakeResult(GA_ERROR::INVALID_PARAMETERS, STR_NONE);
         }
 
         Ride* ride = get_ride(_rideIndex);
         if (ride == nullptr || ride->type == RIDE_TYPE_NULL)
         {
-            log_warning("Invalid game command, ride_id = %u", _rideIndex);
+            log_warning("Invalid game command, ride_id = %u", uint32_t(_rideIndex));
             return MakeResult(GA_ERROR::INVALID_PARAMETERS, STR_NONE);
         }
 
         rct_ride_entry* rideEntry = get_ride_entry(ride->subtype);
         if (rideEntry == nullptr)
         {
-            log_warning("Invalid game command for ride %u", _rideIndex);
+            log_warning("Invalid game command for ride %u", uint32_t(_rideIndex));
             return MakeResult(GA_ERROR::INVALID_PARAMETERS, STR_NONE);
         }
 
@@ -88,14 +88,14 @@ public:
         Ride* ride = get_ride(_rideIndex);
         if (ride == nullptr || ride->type == RIDE_TYPE_NULL)
         {
-            log_warning("Invalid game command, ride_id = %u", _rideIndex);
+            log_warning("Invalid game command, ride_id = %u", uint32_t(_rideIndex));
             return MakeResult(GA_ERROR::INVALID_PARAMETERS, STR_NONE);
         }
 
         rct_ride_entry* rideEntry = get_ride_entry(ride->subtype);
         if (rideEntry == nullptr)
         {
-            log_warning("Invalid game command for ride %u", _rideIndex);
+            log_warning("Invalid game command for ride %u", uint32_t(_rideIndex));
             return MakeResult(GA_ERROR::INVALID_PARAMETERS, STR_NONE);
         }
